Bugfixes and optimizations
Fix the following bugs:

- issue #87: No create user when copying customer account
- invoices, credit memos, reminders and purchase invoices displayed balances as
  -0.00 or mark them red.  Fixed this by rounding payment amount and total
  before computing the balance.
- displayed wrong property (vendor) in show view of purchase invoice
- display errors on small screens in turnover report
- shorten month names on small screens in reports with month selector. Also
  add title attribute with full month name to the links.
- selecting a year in turnover report didn't display the correct report
- wrong variable name in DunningController.workToItem()
- inject UserService in Invoice objects in create action
- fix invalid method call in Google sync class
- improve readability of classes

+++ src/test/groovy/org/amcworld/springcrm/ConfigSpec.groovy
@@ -141,8 +141,8 @@ class ConfigSpec extends Specification {
 
         then: 'I can convert it to an integer type'
         i == c.toType(Integer)
-        i == c.toType(Long)
-        i == c.toType(BigInteger)
+        i as Long == c.toType(Long)
+        i as BigInteger == c.toType(BigInteger)
 
         where:
         v       | i
